Fix username in overseerr SSO logs
Username was being reported as Guest, this is to report the correct username in the logs, was receiving null value from $user['username'] as Plex usernames from Overseerr API uses $user['plexUsername']

Example;

````
/api/v1/auth/plex

{
    "permissions": 2,
    "id": 1,
    "plexUsername": "Username",
    "username": null,
    "recoveryLinkExpirationDate": null,
    "userType": 1,
    "avatar": "https://plex.tv/users/dgerrgrdg/avatar?c=5841564",
    "createdAt": "2021-02-23T22:27:44.000Z",
    "updatedAt": "2021-02-24T13:23:53.000Z",
    "settings": null,
    "requestCount": 1,
    "displayName": "Username",
    "plexToken": "Token",
    "password": null,
    "resetPasswordGuid": null,
    "plexId": null
}
````

Not sure if the intention is to keep the username being reported back from Overseerr, would make sense to make sure that the username is correctly matching I suppose. But everything else is just using the **$username** variable so happy updating PR if this is preferred.